Simple POD adaptation now works. Cleaned up several POD types by putting them in the same file.
